Added comments for Dave
